+++ Electrode/src/Electrode_RadialRegion.cpp
@@ -1724,7 +1724,7 @@ void Electrode_RadialRegion::printElectrode(int pSrc, bool subTimeStep)
         printf("                   time init  = %g\n", t_init_init_);
     }
     printf("   ===============================================================\n");
-    printf("          Number of external surfaces = %d\n", numExternalInterfacialSurfaces_);
+    printf("          Number of external surfaces = %d\n", (int) numExternalInterfacialSurfaces_);
     printf("          Solid Volume = %10.3E\n", ElectrodeSolidVolume_);
     printf("          Total Volume = %10.3E\n", egv);
     printf("          Temperature = %g\n", temperature_);
